src/microservices/events/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
kafka-python==2.0.2
python-dotenv==1.0.1 